llm_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-ollama==0.0.1
requests==2.31.0
python-dotenv==1.0.0

